% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IshwaranJames.R
\name{alphaBoundsIndep}
\alias{alphaBoundsIndep}
\title{Calculate upper bounds for a0 and a1 based on independence assumption}
\usage{
alphaBoundsIndep(p0, p1, draw00, draw01, draw10, draw11)
}
\arguments{
\item{p0}{P(Tobs=1|Z=0)}

\item{p1}{P(Tobs=1|Z=1)}

\item{draw00}{List with mu, p, sigma_sq from DPM for (Tobs=0,Z=0) cell}

\item{draw01}{List with mu, p, sigma_sq from DPM for (Tobs=0,Z=1) cell}

\item{draw10}{List with mu, p, sigma_sq from DPM for (Tobs=1,Z=0) cell}

\item{draw11}{List with mu, p, sigma_sq from DPM for (Tobs=1,Z=1) cell}
}
\value{
List with upper bounds for a0 and a1
}
\description{
Calculate upper bounds for a0 and a1 based on independence assumption
}

