% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{angrist}
\alias{angrist}
\title{Angrist et al. (2002) Dataset}
\format{A data frame with ??? rows and ??? variables:
\describe{
   \item{???}{???}
   \item{???}{???}
   \item{???}{???}
}}
\source{
\url{http://economics.mit.edu/files/1395}
}
\usage{
angrist
}
\description{
Description of the dataset goes here.
}
\references{
\url{http://economics.mit.edu/faculty/angrist/data1/data/angetal02}
}
\keyword{datasets}

