% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDGPs.R
\name{binDGP}
\alias{binDGP}
\title{Simulation DGP for binary example with independent instrument}
\usage{
binDGP(a0, a1, b = 1, c = 0, n = 1000, d = 0.15, rho = 0.5)
}
\arguments{
\item{a0}{Up error rate}

\item{a1}{Down error rate}

\item{b}{ATE}

\item{c}{Intercept}

\item{n}{Sample size}

\item{d}{Controls non-compliance}

\item{rho}{Error correlation}
}
\value{
Samples from model
}
\description{
Simulation DGP for binary example with independent instrument
}
