% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{dinkelman}
\alias{dinkelman}
\title{Dinkelman (2011) Dataset}
\format{A data frame with ??? rows and ??? variables:
\describe{
   \item{???}{???}
   \item{???}{???}
   \item{???}{???}
}}
\source{
\url{https://www.aeaweb.org/aer/data/dec2011/20080791_data.zip}
}
\usage{
dinkelman
}
\description{
Description of the dataset goes here.
}
\references{
\url{http://www.jstor.org/stable/41408731}
}
\keyword{datasets}
