% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{draw_dTstar_tilde_valid}
\alias{draw_dTstar_tilde_valid}
\title{Make draws for dTstar_tilde under the assumption that dz_tilde = 0}
\usage{
draw_dTstar_tilde_valid(y_name, T_name, z_name, controls = NULL, data,
  nRF = 500, nIS = 500)
}
\arguments{
\item{nIS}{}
}
\description{
Make draws for dTstar_tilde under the assumption that dz_tilde = 0
}
\examples{
afghanControls <- c("headchild", "age",  "yrsvill",  "farsi",  "tajik",
                   "farmers",  "agehead",  "educhead",  "nhh",  "land",
                   "sheep", "distschool", "chagcharan")
foo <- draw_dTstar_tilde_valid("testscore", "enrolled", "buildschool",
                     afghanControls, afghan)
}

