% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{draw_dz_tilde}
\alias{draw_dz_tilde}
\title{Make draws for dz_tilde}
\usage{
draw_dz_tilde(y_name, T_name, z_name, controls = NULL, data,
  dTstar_tilde_range, alphas_upper = NULL, nRF = 10, nIS = 5,
  maxIter = nIS * 20)
}
\arguments{
\item{maxIter}{}
}
\description{
Make draws for dz_tilde
}

