% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IshwaranJames.R
\name{drawDPM}
\alias{drawDPM}
\title{Gibbs sampler for Dirichlet Process Mixture of Normals}
\usage{
drawDPM(X, nDraws = 5000, burnIn = 1000, maxClust = 50, A = 1000,
  xi_sq = NULL, eta1 = 2, eta2 = 2, nu1 = 2, nu2 = 2)
}
\arguments{
\item{X}{Vector of input data.}

\item{nDraws}{Number of draws from the Gibbs sampler (exluding burn-in).}

\item{burnIn}{Number of burn-in draws.}

\item{maxClust}{Maximum number of mixture components.}

\item{A}{Prior parameter.}

\item{xi_sq}{Prior parameter.}

\item{eta1}{Prior parameter.}

\item{eta2}{Prior parameter.}

\item{nu1}{Prior parameter.}

\item{nu2}{Prior parameter.}
}
\value{
List of MCMC draws for mu (means of mixture components), sigma_sq
(variances of mixture components), p (mixing probabilities), alpha, theta,
and K (assignment of observations to clusters).
}
\description{
Gibbs sampler for Dirichlet Process Mixture of Normals
}
