% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{drawObs}
\alias{drawObs}
\title{Make draws of reduced form parameters}
\usage{
drawObs(y_name, T_name, z_name, controls = NULL, data, nDraws = 1000)
}
\arguments{
\item{nDraws}{}
}
\value{
Dataframe, each row of which is one draw for the reduced form parameters with column names that match the names of the list generated by the function \code{getObs}
}
\description{
Make draws of reduced form parameters
}
