% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IshwaranJames.R
\name{drawObsIJ}
\alias{drawObsIJ}
\title{Draw reduced form parameters from Ishwaran & James DPM}
\usage{
drawObsIJ(y_name, T_name, z_name, data, nDraws = 1000, burnIn = 1000)
}
\arguments{
\item{y_name}{Name of outcome}

\item{T_name}{Name of treatment}

\item{z_name}{Name of instrument}

\item{data}{Dataframe}

\item{nDraws}{Number of MCMC draws}

\item{burnIn}{Number of burn-in draws}
}
\value{
Matrix of reduced form draws
}
\description{
Draw reduced form parameters from Ishwaran & James DPM
}
