% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{gelbach}
\alias{gelbach}
\title{Gelbach (2002) Dataset}
\format{A data frame with ??? rows and ??? variables:
\describe{
   \item{???}{???}
   \item{???}{???}
   \item{???}{???}
}}
\source{
Provided by author.
}
\usage{
gelbach
}
\description{
Description of the dataset goes here.
}
\references{
\url{http://www.jstor.org/stable/3083335}
}
\keyword{datasets}
