% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_alpha_bounds}
\alias{get_alpha_bounds}
\title{Calculate sharp bounds for a0 and a1}
\usage{
get_alpha_bounds(obs)
}
\arguments{
\item{obs}{List of observables calculated by \code{getObs}.}
}
\value{
Vector with two elements: the first is the upper bound for a0 and the
second is the upper bound for a1. (The lower bounds for each are both zero.)
}
\description{
Calculate sharp bounds for a0 and a1
}
\examples{
afghanControls <- c("headchild", "age",  "yrsvill",  "farsi",  "tajik",
                   "farmers",  "agehead",  "educhead",  "nhh",  "land",
                   "sheep", "distschool", "chagcharan")
afghanY <- "testscore"
afghanT <- "enrolled"
afghanZ <- "buildschool"
afghanObs <- getObs(afghanY, afghanT, afghanZ, afghanControls, afghan)
get_alpha_bounds(afghanObs)
}
