% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_beta}
\alias{get_beta}
\title{Calculate beta}
\usage{
get_beta(dTstar_tilde, a0, a1, obs)
}
\arguments{
\item{dTstar_tilde}{Endogeneity of binary regressor.}

\item{a0}{Probability of observing T = 1 when true treatment status is 0.}

\item{a1}{Probability of observing T = 0 when true treatment status is 1.}

\item{obs}{List of observables calculated by \code{getObs}.}

\item{beta_iv}{Vector of IV coefficients for beta}
}
\value{
Value of beta implied by the data and specified values of regressor endogeneity and mis-classification probabilities.
}
\description{
Calculate beta
}
