% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_dTstar_tilde}
\alias{get_dTstar_tilde}
\title{Calculate endogeneity of binary regressor}
\usage{
get_dTstar_tilde(dz_tilde, a0, a1, obs)
}
\arguments{
\item{dz_tilde}{Endogeneity of binary instrument.}

\item{a0}{Probability of observing T = 1 when true treatment status is 0.}

\item{a1}{Probability of observing T = 0 when true treatment status is 0.}

\item{obs}{List of obvservables calculated using get Obs.}
}
\value{
Endogneity of binary regressor implied by specified values of
instrument invalidity and mis-classification probabilities.
}
\description{
Calculate endogeneity of binary regressor
}

