% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{get_dz_tilde}
\alias{get_dz_tilde}
\title{Calculate endogeneity of binary instrument}
\usage{
get_dz_tilde(dTstar_tilde, a0, a1, obs)
}
\arguments{
\item{dTstar_tilde}{Endogeneity of binary regressor.}

\item{a0}{Probability of observing T = 1 when true treatment status is 0.}

\item{a1}{Probability of observing T = 0 when true treatment status is 1.}

\item{obs}{List of observables calculated by \code{getObs}.}
}
\value{
Value of instrument endogeneity implied by the data and specified values of regressor endogeneity and mis-classification probabilities.
}
\description{
Calculate endogeneity of binary instrument
}

