% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_dz_tilde_check}
\alias{get_dz_tilde_check}
\title{If constraint is satisfied, calculate endogneity of binary instrument}
\usage{
get_dz_tilde_check(dTstar_tilde, a0, a1, obs)
}
\arguments{
\item{dTstar_tilde}{Endogeneity of binary regressor.}

\item{a0}{Probability of observing T = 1 when true treatment status is 0.}

\item{a1}{Probability of observing T = 0 when true treatment status is 1.}

\item{obs}{List of observables calculated by \code{getObs}.}
}
\value{
If the variance constraint is satistfied, return the value of instrument endogeneity implied by the data and specified values of instrument endogeneity and mis-classification probabilities. Otherwise return \code{NA}.
}
\description{
If constraint is satisfied, calculate endogneity of binary instrument
}
\examples{
AngristControls <- c("age", "svy", "sex2", "phone", "hsvisit",
                     "d1995", "djamundi", "dmonth1", "dmonth2",
                     "dmonth3", "dmonth4", "dmonth5", "dmonth6", "dmonth7",
                     "dmonth8", "dmonth9", "dmonth10", "dmonth11",
                     "strata1", "strata2", "strata3", "strata4", "strata5")
AngristY <- "totalRepeats"
AngristT <- "usesch"
AngristZ <- "vouch0"
AngristObs <- getObs(AngristY, AngristT, AngristZ, AngristControls, angrist)
get_dz_tilde_check(-0.2, 0.1, 0.1, AngristObs)
get_dz_tilde_check(-0.1, 0.1, 0.2, AngristObs)
}

