% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{get_m11_bounds}
\alias{get_m11_bounds}
\title{Bounds on m11 from variance conditions}
\usage{
get_m11_bounds(a0, a1, obs)
}
\arguments{
\item{a0}{Probability of observing T = 1 when true treatment status is 0.}

\item{a1}{Probability of observing T = 0 when true treatment status is 1.}

\item{obs}{List of observables calculated using \code{getObs}.}
}
\value{
A list of bounds for m11 or NA if there are no bounds.
}
\description{
Bounds on m11 from variance conditions
}
\examples{
noCov <- getObs("totalRepeats", "usesch", "vouch0", data = angrist)
get_m11_bounds(0.2, 0.4, noCov)
}

