% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{getMomentsForSampler}
\alias{getMomentsForSampler}
\title{Get moments for sampling reduced-form parameters}
\usage{
getMomentsForSampler(y_name, T_name, z_name, controls = NULL, data)
}
\arguments{
\item{data}{}
}
\description{
Get moments for sampling reduced-form parameters
}
