% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{getObs}
\alias{getObs}
\title{Calculate summary statistics of observables}
\usage{
getObs(y_name, T_name, z_name, controls = NULL, data)
}
\arguments{
\item{y_name}{Name of the outcome variable.}

\item{T_name}{Name of the potentially endogenous and mis-measured binary
treatment variable.}

\item{z_name}{Name of the potentially endogenous binary instrument.}

\item{controls}{Vector of names of exogenous control regressors.}

\item{data}{Name of dataframe whose columns contain the data for the outcome,
regressor, instrument, and controls.}
}
\value{
A list of summary statistics: probabilities, conditional means and
variances, etc., in notation corresponding to the paper.
}
\description{
This function calculates summary statistics for later use in the
partial identification exercise for a binary, mis-measured, endogenous
regressor based on a binary instrument that may itsef be endogenous.
}
\examples{
AngristControls <- c("age", "svy", "sex2", "phone", "hsvisit",
                     "d1995", "djamundi", "dmonth1", "dmonth2",
                     "dmonth3", "dmonth4", "dmonth5", "dmonth6", "dmonth7",
                     "dmonth8", "dmonth9", "dmonth10", "dmonth11",
                     "strata1", "strata2", "strata3", "strata4", "strata5")
AngristY <- "totalRepeats"
AngristT <- "usesch"
AngristZ <- "vouch0"
getObs(AngristY, AngristT, AngristZ, AngristControls, angrist)

}

