% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IshwaranJames.R
\name{onlyActive}
\alias{onlyActive}
\title{Process results of drawDPM to drop empty clusters}
\usage{
onlyActive(draws)
}
\arguments{
\item{draws}{Output of drawDPM}
}
\value{
List of lists with values for non-empty clusters
}
\description{
Process results of drawDPM to drop empty clusters
}

