% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_fixed_a1}
\alias{plot_fixed_a1}
\title{Plot fixing a1}
\usage{
plot_fixed_a1(obs, a1 = 0, dTstar_lim, n_grid = 50, theta = 0, phi = 15,
  forceZero = FALSE, xlab = "a0", ylab = "dTstar", zlab = "dz",
  col = "dodgerblue2")
}
\arguments{
\item{col}{}
}
\description{
Plot fixing a1
}
