% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_fixed_dTstar}
\alias{plot_fixed_dTstar}
\title{Fixed dTstar plot}
\usage{
plot_fixed_dTstar(obs, dTstar, a0_range = seq(0, 0.2, 0.01),
  a1_range = seq(0, 0.2, 0.01), theta = 0, phi = 15, forceZero = TRUE)
}
\arguments{
\item{forceZero}{}
}
\description{
Fixed dTstar plot
}

