% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCDFs.R
\name{plotAlphaBounds}
\alias{plotAlphaBounds}
\title{Make plot of bounds on a0 and a1 implied by observable CDFs}
\usage{
plotAlphaBounds(inData, inc = 0.01, j = 10)
}
\arguments{
\item{inData}{Dataframe with columns named Tobs (binary), y, and z (binary)}

\item{inc}{Increment for tau}

\item{frac}{Location of dashed lines: quantiles of y distribution}
}
\value{
Make plot of bounds on a0 and a1
}
\description{
Make plot of bounds on a0 and a1 implied by observable CDFs
}
