% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_dTstar_draws}
\alias{summarize_dTstar_draws}
\title{Summarize posterior draws from draw_dTstar_tilde}
\usage{
summarize_dTstar_draws(draws)
}
\arguments{
\item{draws}{}
}
\description{
Summarize posterior draws from draw_dTstar_tilde
}

