% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_dz_draws}
\alias{summarize_dz_draws}
\title{Summarize posterior draws from draw_dz_tilde}
\usage{
summarize_dz_draws(draws, digits = 2)
}
\description{
Summarize posterior draws from draw_dz_tilde
}

