% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IshwaranJames.R
\name{cellOutput}
\alias{cellOutput}
\title{Generate reduced form draws for particular cell}
\usage{
cellOutput(y_cell)
}
\arguments{
\item{y_cell}{Observations for one of the four cells.}
}
\value{
Reduced form draws
}
\description{
Generate reduced form draws for particular cell
}

