% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_fixed_dTstar}
\alias{plot_fixed_dTstar}
\title{Fixed dTstar plot}
\usage{
plot_fixed_dTstar(obs, dTstar, alpha_grid = 0.005, theta = 0, phi = 15,
  forceZero = TRUE, xlab = "a0", ylab = "a1", zlab = "dz",
  col = "dodgerblue2")
}
\arguments{
\item{forceZero}{}
}
\description{
Fixed dTstar plot
}
